# Conflicts:
#   backend/src/main/java/com/hankki/common/gpt/GptPromptBuilder.java
#   backend/src/main/java/com/hankki/domain/recommend/service/RecommendFacade.java
#   backend/src/main/java/com/hankki/domain/recommend/service/RecommendVectorFacade.java

+++ backend/src/main/java/com/hankki/domain/food/dto/FoodPreviewResponseDto.java
@@ -28,5 +28,5 @@ public class FoodPreviewResponseDto {
 
     @NotBlank
     @Schema(description = "1인분 기준 칼로리", example = "300")
-    private int kcal;
+    private double kcal;
 }
